Moved utility classes to their own files
