+++ client/src/modules/core/utils/fileutils.js
@@ -1,0 +1,115 @@
+/**
+ * BetterDiscord Client Utils Module
+ * Copyright (c) 2015-present JsSucks - https://github.com/JsSucks
+ * All rights reserved.
+ * https://github.com/JsSucks - https://betterdiscord.net
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+const fs = window.require('fs');
+
+class FileUtils {
+
+    static async fileExists(path) {
+        return new Promise((resolve, reject) => {
+            fs.stat(path, (err, stats) => {
+                if (err) return reject({
+                    'message': `No such file or directory: ${err.path}`,
+                    err
+                });
+
+                if (!stats.isFile()) return reject({
+                    'message': `Not a file: ${path}`,
+                    stats
+                });
+
+                resolve();
+            });
+        });
+    }
+
+    static async directoryExists(path) {
+        return new Promise((resolve, reject) => {
+            fs.stat(path, (err, stats) => {
+                if (err) return reject({
+                    'message': `Directory does not exist: ${path}`,
+                    err
+                });
+
+                if (!stats.isDirectory()) return reject({
+                    'message': `Not a directory: ${path}`,
+                    stats
+                });
+
+                resolve();
+            });
+        });
+    }
+
+    static async readFile(path) {
+        try {
+            await this.fileExists(path);
+        } catch (err) {
+            throw (err);
+        }
+
+        return new Promise((resolve, reject) => {
+            fs.readFile(path, 'utf-8', (err, data) => {
+                if (err) reject({
+                    'message': `Could not read file: ${path}`,
+                    err
+                });
+
+                resolve(data);
+            });
+        });
+    }
+
+    static async writeFile(path, data) {
+        return new Promise((resolve, reject) => {
+            fs.writeFile(path, data, err => {
+                if (err) return reject(err);
+                resolve();
+            });
+        });
+    }
+
+    static async readJsonFromFile(path) {
+        let readFile;
+        try {
+            readFile = await this.readFile(path);
+        } catch (err) {
+            throw (err);
+        }
+
+        try {
+            const parsed = await Utils.tryParseJson(readFile);
+            return parsed;
+        } catch (err) {
+            throw (Object.assign(err, { path }));
+        }
+    }
+
+    static async writeJsonToFile(path, json) {
+        return this.writeFile(path, JSON.stringify(json));
+    }
+
+    static async readDir(path) {
+        try {
+            await this.directoryExists(path);
+            return new Promise((resolve, reject) => {
+                fs.readdir(path, (err, files) => {
+                    if (err) return reject(err);
+                    resolve(files);
+                });
+            });
+        } catch (err) {
+            throw err;
+        }
+    }
+
+}
+
+module.exports = { FileUtils };
