+++ client/src/modules/core/utils/utils.js
@@ -1,0 +1,34 @@
+/**
+ * BetterDiscord Client Utils Module
+ * Copyright (c) 2015-present JsSucks - https://github.com/JsSucks
+ * All rights reserved.
+ * https://github.com/JsSucks - https://betterdiscord.net
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+class Utils {
+
+    static overload(fn, cb) {
+        const orig = fn;
+        return function(...args) {
+            orig(...args);
+            cb(...args);
+        }
+    }
+
+    static async tryParseJson(jsonString) {
+        try {
+            return JSON.parse(jsonString);
+        } catch (err) {
+            throw ({
+                'message': 'Failed to parse json',
+                err
+            });
+        }
+    }
+
+}
+
+module.exports = { Utils };
